Seems strange (and the code backs up) that if the max and min of a statistic is expressed as a double, the last value would not also need to be a double.
(closes issue #15807)
 Reported by: klaus3000


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@268773 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ include/asterisk/rtp_engine.h
@@ -235,9 +235,9 @@ struct ast_rtp_instance_stats {
 	/*! Number of packets received */
 	unsigned int rxcount;
 	/*! Jitter on transmitted packets */
-	unsigned int txjitter;
+	double txjitter;
 	/*! Jitter on received packets */
-	unsigned int rxjitter;
+	double rxjitter;
 	/*! Maximum jitter on remote side */
 	double remote_maxjitter;
 	/*! Minimum jitter on remote side */
@@ -275,7 +275,7 @@ struct ast_rtp_instance_stats {
 	/*! Standard deviation packets lost on local side */
 	double local_stdevrxploss;
 	/*! Total round trip time */
-	unsigned int rtt;
+	double rtt;
 	/*! Maximum round trip time */
 	double maxrtt;
 	/*! Minimum round trip time */
